rename toggleTheme service to just theme
